src/requirements.txt
tensorflow==2.13.0
keras==2.13.1
mlflow==2.4.1
pandas
pillow==10.0.0
opencv-python
scikit-learn==1.4.0